Conflicts:
	opencga-storage/opencga-storage-core/src/main/java/org/opencb/opencga/storage/core/variant/adaptors/VariantDBAdaptor.java

+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb/converters/CohortConverter.java
@@ -34,7 +34,7 @@ public class CohortConverter extends GenericConverter<Cohort, Document> {
         Document document = null;
         try {
             document = Document.parse(CohortWriter.writeValueAsString(object));
-            document.put("id", document.getLong("id"));
+            document.put("id", document.getInteger("id").longValue());
         } catch (JsonProcessingException e) {
             e.printStackTrace();
         }
