+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb/converters/IndividualConverter.java
@@ -35,7 +35,7 @@ public class IndividualConverter extends GenericConverter<Individual, Document> 
         Document document = null;
         try {
             document = Document.parse(individualWriter.writeValueAsString(object));
-            document.put("id", document.getLong("id"));
+            document.put("id", document.getInteger("id").longValue());
         } catch (JsonProcessingException e) {
             e.printStackTrace();
         }
