+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb/converters/VariableSetConverter.java
@@ -34,7 +34,7 @@ public class VariableSetConverter extends GenericConverter<VariableSet, Document
         Document document = null;
         try {
             document = Document.parse(variableSetWriter.writeValueAsString(object));
-            document.put("id", document.getLong("id"));
+            document.put("id", document.getInteger("id").longValue());
         } catch (JsonProcessingException e) {
             e.printStackTrace();
         }
